fix top up view back for sale widget
