+++ packages/checkout/widgets-lib/src/widgets/sale/views/PaymentMethods.tsx
@@ -20,10 +20,8 @@ import { PaymentOptions } from '../components/PaymentOptions';
 import { useSaleContext } from '../context/SaleContextProvider';
 import { useSaleEvent } from '../hooks/useSaleEvents';
 import { SaleErrorTypes, SignPaymentTypes } from '../types';
-import { useInsufficientBalance } from '../hooks/useInsufficientBalance';
 
 export function PaymentMethods() {
-  useInsufficientBalance();
   const { t } = useTranslation();
   const { viewDispatch } = useContext(ViewContext);
   const {
